pyproject.toml
[project]
name = "chatbot-webui-backend"
version = "0.1.0"
description = "Backend services for the Chatbot Web UI v2"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Project Team" }]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.21",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "apscheduler>=3.10.4",
    "structlog>=24.1.0",
    "fastapi-sse>=1.0.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.1",
    "cryptography>=43.0.0"
]

[project.optional-dependencies]
test = ["pytest>=8.3.0", "pytest-asyncio>=0.23.7"]
lint = ["ruff>=0.5.6", "mypy>=1.10.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api"]

[tool.hatch.build.targets.sdist]
include = ["api", "tests", "pyproject.toml", "README.md"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
plugins = []
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
